% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gardon.R
\docType{data}
\name{cours_eau}
\alias{cours_eau}
\title{Couche cartographique des cours d'eau en Ile de France}
\format{
\subsection{\code{cours_eau}}{

Un sf object avec 2031 observations et 16 variables:
\describe{
\item{ID}
\item{CODE_HYDRO}
\item{TOPONYME}
\item{STATUT_TOP}
\item{IMPORTANCE}
\item{DATE_CREAT}
\item{DATE_MAJ}
\item{DATE_APP}
\item{DATE_CONF}
\item{SOURCE}
\item{ID_SOURCE}
\item{STATUT}
\item{MAREE}
\item{PERMANENT}
\item{COMMENT}
\item{geometry}
}
}
}
\usage{
cours_eau
}
\description{
Couche cartographique des cours d'eau en Ile de France
}
\keyword{datasets}
