% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espece.R
\name{distri_sp}
\alias{distri_sp}
\title{Distribution des especes}
\usage{
distri_sp(
  data,
  dir = getwd(),
  folder = NULL,
  file.name = "distri_sp",
  save = FALSE,
  boxplot = TRUE,
  map = TRUE,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{folder}{Dossier ou enregistrer les cartes et les boxplots. Si \code{folder = NULL}, un nouveau dossier ("distribution_sp") est cree dans \code{dir}.}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "distri_sp"}.}

\item{save}{Argument logique \code{TRUE/FALSE}. Par defaut \code{save = FALSE} et les graphiques produits ne seront pas enregistres dans un repertoire local.}

\item{boxplot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un boxplot des effectifs aux stations par espece est trace.}

\item{map}{Argument logique \code{TRUE/FALSE}. Par defaut \code{map = TRUE} et une carte des alertes est tracee.}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}
}
\value{
Des resultats graphiques : carte (\code{map = TRUE}) et boxplot (\code{boxplot = TRUE}).
}
\description{
Fonction de visualisation des effectifs par stations (boxplot) et de la distribution des especes sur la zone concernees.
}
\examples{
result_distri <- distri_sp(
data = exemple_IDF,
save = FALSE,
boxplot = TRUE,
map = TRUE)
}
\seealso{
Une fonction de visualisation de la frequence d'apparition (occurrence) des especes dans les operations : \code{\link{test_sp_occ}}.
}
