% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{make_tab}
\alias{make_tab}
\title{Creation de tableaux au sein des fonctions}
\usage{
make_tab(
  x,
  colours = NULL,
  titre,
  comment,
  size.tab = 10,
  size.titre = 12,
  comment.size = 9
)
}
\arguments{
\item{x}{Data au format dataframe.}

\item{colours}{Couleurs pour chaque cellules du tableau en partant de la gauche.}

\item{titre}{Titre du tableau entre \code{""}.}

\item{comment}{Character entre \code{""}. Commentaire a ajouter en dessous du tableau.}

\item{size.tab}{Numerique. Taille des caracteres du tableau.}

\item{size.titre}{Numerique. Taille du titre.}

\item{comment.size}{Numerique. Taille du commentaire.}
}
\value{
Un tableau contenant les statistiques de la fonction.
}
\description{
Permet de creer des tableaux en modifiant la taille de la police, les couleurs, le titre, le commentaire.
}
\examples{
table_stat <- make_tab(x = tempJanv_stat,
                       colours = c("#FF6BA1", "#FF6BA1",
                                   "#FF6BA1", "#73FF7D"),
                       titre = "Rapport Temperature janvier (\%)",
                       comment = commentaire)

}
