% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espece.R
\name{Occ_Annees}
\alias{Occ_Annees}
\title{Occurrences des especes par an}
\usage{
Occ_Annees(data, dir = getwd(), folder = NULL, save = FALSE, plot = TRUE)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{folder}{Dossier ou enregistrer les cartes et les boxplots. Si \code{folder = NULL}, un nouveau dossier ("Occ_Annees") est cree dans \code{dir}.}

\item{save}{Argument logique \code{TRUE/FALSE}. Par defaut \code{save = FALSE} et les graphiques produits ne seront pas enregistres dans un repertoire local.}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un graphique des occurrences par espece est trace.}
}
\value{
Des resultats graphiques si \code{plot = TRUE}.
}
\description{
Fonction de visualisation de la frequence d'apparition (Occurrence) des especes au cours du temps.
}
\examples{
occurrence_annees <- Occ_Annees(
data = IDF_new,
plot = TRUE)
}
