% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espece.R
\name{pop_diversite}
\alias{pop_diversite}
\title{Richesse specifique aux points de prelevement}
\usage{
pop_diversite(data, dir = getwd(), folder = NULL, save = FALSE, plot = TRUE)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{folder}{Dossier ou enregistrer les cartes et les boxplots. Si \code{folder = NULL}, un nouveau dossier ("distribution_sp") est cree dans \code{dir}.}

\item{save}{Argument logique \code{TRUE/FALSE}. Par defaut \code{save = FALSE} et les graphiques produits ne seront pas enregistres dans un repertoire local.}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un graphique de la richesse specifique au cours du temps pour chaque point de prelevement est trace.}
}
\value{
Des resultats graphiques si \code{plot = TRUE}.
}
\description{
Fonction de visualisation de la richesse specifique aux points de prelevements au cours du temps.
}
\examples{
diversite <- pop_diversite(
data = IDF_new,
plot = TRUE)
}
