% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_operation.R
\name{test_ope_CalSurf}
\alias{test_ope_CalSurf}
\title{Verification du calcul de surface de l'operation}
\usage{
test_ope_CalSurf(
  data,
  plot = TRUE,
  stat = TRUE,
  dir = getwd(),
  file.name = "ope_CalSurf",
  map = TRUE,
  csv = TRUE,
  color_alerte = pal_alerte,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau,
  pdf = TRUE
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un boxplot de la dispertion des valeurs est trace.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "ope_CalSurf"}.}

\item{map}{Argument logique \code{TRUE/FALSE}. Par defaut \code{map = TRUE} et une carte des alertes est tracee.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}

\item{color.alerte}{Vecteur. Contient les deux couleurs indiquant une alerte ou non. Par defaut, \code{color_alerte = pal_alerte}.}
}
\value{
Des resultats graphiques : boxplot (\code{plot = TRUE}), carte (\code{map = TRUE}) et tableaux (\code{stat = TRUE}). Un pdf reprenant ces precedents resultats. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Test qui permet de verifier la conformite du calcul de la surface en fonction du protocole d'echantillonnage.
}
\examples{
result_CalSurf <- test_ope_CalSurf(
data = exemple_IDF,
pdf = FALSE,
csv = FALSE)
}
\seealso{
Pour verifier la coherence des valeurs de surface de l'operation : \code{\link{test_ope_surf}}.
}
