% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lots.R
\name{test_ope_lotTyl}
\alias{test_ope_lotTyl}
\title{Conformite des lots}
\usage{
test_ope_lotTyl(
  data,
  stat = TRUE,
  dir = getwd(),
  file.name = "ope_lotTyl",
  map = TRUE,
  csv = TRUE,
  color.point = colorblind_8,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau,
  pdf = TRUE
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "ope_lotTyl"}.}

\item{map}{Argument logique \code{TRUE/FALSE}. Par defaut \code{map = TRUE} et une carte des alertes est tracee.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{color.point}{Vecteur. Palette contenant les couleurs pour les points de la carte. Par defaut, \code{color.point = colorblind_8}.}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}
}
\value{
Des resultats graphiques : carte (\code{map = TRUE}) et tableaux (\code{stat = TRUE}). Un pdf reprenant ces precedents resultats. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Verifie que les caracteristiques de chaque type lot sont conforme (effectif, poids, taille)
}
\examples{
result_lotTyl <- test_ope_lotTyl(
data = exemple_IDF,
csv = FALSE,
pdf = FALSE
)
}
