% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocole.R
\name{test_ope_mesure}
\alias{test_ope_mesure}
\title{Coherence des valeurs de profondeur, largeur et longueur.}
\usage{
test_ope_mesure(
  data,
  stat = TRUE,
  dir = getwd(),
  plot = TRUE,
  file.name = "ope_mesure",
  csv = TRUE,
  color_alerte = pal_alerte,
  pdf = TRUE,
  percent = 0
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un boxplot des ecarts de longeur est trace.}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "ope_mesure"}.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{color_alerte}{Vecteur. Contient les deux couleurs indiquant une alerte ou non. Par defaut, \code{color_alerte = pal_alerte}.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}

\item{percent}{Numerique. Pourcentage d'ecart a la valeur theorique de longueur calculee.}
}
\value{
Des resultats graphiques : boxplot (\code{plot = TRUE}) et tableaux (\code{stat = TRUE}). Un pdf reprenant ces precedents resultats. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Test qui (1) calcule une longueur theorique minimum de l'operation en fonction de la largeur, (2) determine si le protocole de peche est adapte en fonction de la largeur et de la longueur de l'operation, et (3) compare la valeur de longueur theorique à celle realisee.
}
\examples{
result_mesure <- test_ope_mesure(
data = exemple_IDF,
csv = FALSE,
pdf = FALSE)

}
\seealso{
Attention ! Pour utiliser cette fonction, il est preferable de verifier les valeurs de profondeur (\code{\link{test_ope_profondeur}}), de largeur (\code{\link{test_ope_Lar}})et de longueur (\code{\link{test_ope_Lon}}).
}
