% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profondeur.R
\name{test_ope_ProfFacies}
\alias{test_ope_ProfFacies}
\title{Verification du calcul de la profondeur}
\usage{
test_ope_ProfFacies(
  data,
  stat = TRUE,
  dir = getwd(),
  file.name = "ope_ProfFacies",
  csv = TRUE,
  pdf = TRUE
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "ope_ProfFacies"}.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}
}
\value{
Des resultats sous forme de tableaux (\code{stat = TRUE}). Un pdf reprenant ces tableaux. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Verifie que la profondeur des peches partielles correspond bien à la somme des profondeurs des facies.
}
\examples{
result_ProfFacies <- test_ope_ProfFacies(
data = exemple_IDF,
pdf = FALSE,
csv = FALSE)

}
\seealso{
Une fonction permettant de verifier la coherence des valeurs de profondeur : \code{\link{test_ope_profondeur}}. Une fonction exploitant la relation entre la largeur, la profondeur et la longueur de l'operation : \code{\link{test_ope_mesure}}.
}
