% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altitude.R
\name{test_pop_altIDF}
\alias{test_pop_altIDF}
\title{Coherence des valeurs d'altitude par rapport a un modele.}
\usage{
test_pop_altIDF(
  data,
  modele = altitude_IDF,
  plot = TRUE,
  stat = TRUE,
  dir = getwd(),
  file.name = "pop_altIDF",
  map = TRUE,
  csv = TRUE,
  method = 1,
  echelle = "A",
  color.alerte = pal_alerte,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau,
  pdf = TRUE
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{modele}{Modele d'altitude locale a utiliser. Par defaut, \code{modele = altitude_IDF} un modele uniquement pour l'Ile de France.}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un boxplot de la dispertion des valeurs est trace.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "pop_altIDF"}.}

\item{map}{Argument logique \code{TRUE/FALSE}. Par defaut \code{map = TRUE} et une carte des alertes est tracee.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{method}{Une methode pour identifier les outliers. Si \code{method = 1}, les valeurs en dehors de l'interval defini par le \code{min} et le \code{max} sont considerees comme des outliers.}

\item{echelle}{Argument character. Prend deux valeurs possibles \code{"A"/"B"}. Si \code{echelle = "A"}, identifie les outliers a l'echelle des masses d'eau. Si \code{echelle = "B"}, identifie les outliers a l'echelle des points de prelevements.}

\item{color.alerte}{Vecteur. Contient les deux couleurs indiquant une alerte ou non. Par defaut, \code{color_alerte = pal_alerte}.}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}
}
\value{
Des resultats graphiques : boxplot (\code{plot = TRUE}), carte (\code{map = TRUE}) et tableaux (\code{stat = TRUE}). Un pdf reprenant ces precedents resultats. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Deux methodes et deux echelles disponibles afin de tester la coherence des valeurs d'altitude par rapport a un modele d'altitude d'IDF.
}
\examples{
test_pop_altIDF(
data = exemple_IDF,
modele = altitude_IDF,
pdf = FALSE,
csv = FALSE,
echelle = "B")
}
\seealso{
En dehors de l'Ile de France ou sur des stations differentes, une fonction permettant de verifier la coherence des valeurs d'altitude : \code{\link{test_pop_alt}}.
}
