% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_source.R
\name{test_pop_DS}
\alias{test_pop_DS}
\title{Verification relation entre distance_source, surface bassin versant et altitude}
\usage{
test_pop_DS(
  data,
  stat = TRUE,
  dir = getwd(),
  file.name = "pop_DS",
  map = TRUE,
  csv = TRUE,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau,
  pdf = TRUE
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "pop_DS"}.}

\item{map}{Argument logique \code{TRUE/FALSE}. Par defaut \code{map = TRUE} et une carte des alertes est tracee.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}
}
\description{
Fonction qui verifie si la surface du bassin versant augmente et l'altitude diminue avec l'augmentation de la distance a la source.
}
\examples{
result_DS <- test_pop_DS(
data = exemple_IDF,
pdf = FALSE,
csv = FALSE
)
}
\seealso{
Pour verifier la coherence des valeurs de distance a la source : \code{\link{test_pop_dist}}. Pour verifier la coherence des valeurs de surface de bassin versant : \code{\link{test_pop_surfBV}}. Pour verifier la coherence des valeurs d'altitude : \code{\link{test_pop_alt}}.
}
