% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_BV.R
\name{test_pop_surfBV}
\alias{test_pop_surfBV}
\title{Coherence des valeurs de surface de bassin versant}
\usage{
test_pop_surfBV(
  data,
  plot = TRUE,
  stat = TRUE,
  min = 100,
  max = 2000,
  dir = getwd(),
  file.name = "pop_surfBV",
  map = TRUE,
  csv = TRUE,
  method = 2,
  confidence.interval = 95,
  color.alerte = pal_alerte,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau,
  pdf = TRUE,
  percent = 100
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un boxplot de la dispertion des valeurs est trace.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{min}{Valeur minimum a choisir si \code{method = 1}.}

\item{max}{Valeur maximum a choisir si \code{method = 1}.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "pop_surfBV"}.}

\item{map}{Argument logique \code{TRUE/FALSE}. Par defaut \code{map = TRUE} et une carte des alertes est tracee.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{method}{Quatre methodes pour identifier les outliers. Si \code{method = 1}, les valeurs en dehors de l'interval defini par le \code{min} et le \code{max} sont considerees comme des outliers. Si \code{method = 2}, les valeurs en dehors de l'interval defini par l'IQR criterion sont considerees comme des outliers. Si \code{method = 3}, les valeurs en dehors de l'interval defini par un interval de confiance (\code{confidence.interval}) sont considerees comme des outliers. Si \code{method = 4}, une courbe de tendance entre la surface de bassin versant et la distance a la source est tracee et les outliers sont identifie selon un pourcentage d'ecart a cette tendance.}

\item{confidence.interval}{Numerique. Interval de confiance pris en compte si \code{method = 3}. Par defaut \code{confidence.interval = 95}.}

\item{color.alerte}{Vecteur. Contient les deux couleurs indiquant une alerte ou non. Par defaut, \code{color_alerte = pal_alerte}.}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}

\item{percent}{Pourcentage d'ecart a la tendance a prendre en compte si \code{method = 4} et pourcentage d'ecart aux valeurs seuils si \code{method = 2}.}
}
\value{
Des resultats graphiques : boxplot (\code{plot = TRUE}), carte (\code{map = TRUE}) et tableaux (\code{stat = TRUE}). Un pdf reprenant ces precedents resultats. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Quatre methodes disponibles afin de tester la coherence des valeurs surface de bassin versant.
}
\details{
En ce qui concerne l'IQR criterion (\code{method = 2}), les valeurs considerees comme aberrantes sont celles qui sont superieures a \code{quantile(x, 3/4) + IQR(x)} et celles qui sont inferieures a \code{quantile(x, 1/4) - IQR(x)}. Pour plus de details sur le calcul voir les fonctions suivantes : \code{\link[stats]{IQR}} et \code{\link[stats]{quantile}}.
}
\examples{
result_surfBv <- test_pop_surfBV(
data = exemple_IDF,
pdf = FALSE,
csv = FALSE,
method = 4,
percent = 100)
}
\seealso{
Pour verifier la que les valeurs de surface de bassin versant augmente bien avec l'augmentation de la distance a la source : \code{\link{test_pop_DS}}.
}
