% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lots.R
\name{test_sp_effectif}
\alias{test_sp_effectif}
\title{Conformite des effectifs par espece}
\usage{
test_sp_effectif(
  data,
  stat = TRUE,
  dir = getwd(),
  file.name = "sp_effectif",
  plot = TRUE,
  map = TRUE,
  save = FALSE,
  csv = TRUE,
  folder = NULL,
  color.alerte = pal_alerte,
  pdf = TRUE,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau,
  echelle = "espece"
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "ope_lotTyl"}.}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un boxplot de la dispertion des valeurs d'effectif est trace.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{color.alerte}{Vecteur. Palette contenant les deux couleurs des alertes. Par defaut, \code{color.point = pal_alerte}.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}

\item{echelle}{Deux valeurs possibles. Si \code{echelle = espece} alors les effectifs sont evalues par espece sans distinction de station. Si \code{echelle = POP} alors les effectifs sont evalues par espece et par point de prelevement.}

\item{percent}{Numerique. Pourcentage de difference au maximum a partir duquel on identifie les outliers.}
}
\value{
Des resultats graphiques : boxplot (\code{plot = TRUE}) et tableaux (\code{stat = TRUE}). Un pdf reprenant ces precedents resultats. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Identifie les effectifs aberrants pour chaque espece a l'aide de l'IQR criterion.
}
\details{
En ce qui concerne l'IQR criterion (\code{method = 2}), les valeurs considerees comme aberrantes sont celles qui sont superieures a \code{quantile(x, 3/4) + IQR(x)} et celles qui sont inferieures a \code{quantile(x, 1/4) - IQR(x)}. Pour plus de details sur le calcul voir les fonctions suivantes : \code{\link[stats]{IQR}} et \code{\link[stats]{quantile}}.
}
\examples{
result_sp_effectif <- test_sp_effectif(
data = exemple_IDF,
csv = FALSE,
pdf = FALSE,
percent = 100,
method = 2)
}
