% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espece.R
\name{test_sp_occ}
\alias{test_sp_occ}
\title{Occurence des especes}
\usage{
test_sp_occ(
  data,
  stat = TRUE,
  dir = getwd(),
  file.name = "sp_occ",
  plot = TRUE,
  map = TRUE,
  csv = TRUE,
  color_point = colorblind_8,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau,
  pdf = TRUE
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "sp_occ"}.}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un boxplot du pourcentage d'occurrence de chaque espece est trace.}

\item{map}{Argument logique \code{TRUE/FALSE}. Par defaut \code{map = TRUE} et une carte des alertes est tracee.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}

\item{color.point}{Vecteur. Palette contenant les couleurs pour les differentes categories d'occurrence (espece commune, espece peu commune, espece rare et espece non renseignee). Par defaut, \code{color.point = colorblind_8}.}
}
\value{
Des resultats graphiques : carte (\code{map = TRUE}) et tableaux (\code{stat = TRUE}). Un pdf reprenant ces precedents resultats. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Fonction de visualisation de la frequence d'apparition (occurrence) des especes dans les operations.
}
\examples{
result_occ <- test_sp_occ(
data = exemple_IDF,
pdf = FALSE,
csv = FALSE
)
}
\seealso{
Une fonction de visualisation de la distribution spatiale et des effectifs par espece : \code{\link{distri_sp}}.
}
