% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nombre_pop_station.R
\name{test_sta_nombrePop}
\alias{test_sta_nombrePop}
\title{Nombre de point de prelevement par station}
\usage{
test_sta_nombrePop(
  data,
  stat = TRUE,
  dir = getwd(),
  file.name = "sta_nbPop",
  map = TRUE,
  csv = TRUE,
  color.point = divergent_24,
  map.contour = IDF_contour,
  map.cours.eau = cours_eau,
  pdf = TRUE
)
}
\arguments{
\item{data}{Donnees de sortie de la fonction \code{\link{getDataRegion}}. Ou donnees en exemple : \code{data = exemple_IDF}.}

\item{stat}{Argument logique \code{TRUE/FALSE}. Par defaut \code{stat = TRUE} et deux tableaux contenants les resultats de la fonction sont affiches.}

\item{dir}{Chemin d'acces ou enregistrer les fichiers de sortie. Par defaut regle sur le repertoire de travail (\verb{dir = }\code{\link[base]{getwd}}).}

\item{file.name}{Nom a donner aux fichiers en sortie. Par defaut \code{file.name = "sta_nbPop"}.}

\item{map}{Argument logique \code{TRUE/FALSE}. Par defaut \code{map = TRUE} et une carte des alertes est tracee.}

\item{csv}{Argument logique \code{TRUE/FALSE}. Par defaut \code{csv = TRUE} et le fichier csv est enregistre dans le repertoire de travail ou le repertoire indique (\code{dir}).}

\item{color.point}{Vecteur. Contient les couleurs des points sur la carte entre \code{""}. Par defaut, \code{color.point = divergent_24}.}

\item{map.contour}{Format sf. Donnees de contour de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.contour = IDF_contour} qui correspond au fond de carte d'Ile de France.}

\item{map.cours.eau}{Format sf. Donnees des cours d'eau de la zone etudiee a charger avec la fonction \code{\link[sf]{st_read}}. Par defaut, \code{map.cours.eau = cours_eau} qui correspond aux cours d'eau Ile de France.}

\item{pdf}{Argument logique \code{TRUE/FALSE}. Par defaut \code{pdf = TRUE} et un pdf avec les figures est cree dans le repertoire indique (\code{dir}).}

\item{plot}{Argument logique \code{TRUE/FALSE}. Par defaut \code{plot = TRUE} et un boxplot de la dispertion des valeurs est trace.}
}
\value{
Des resultats graphiques : carte (\code{map = TRUE}) et tableaux (\code{stat = TRUE}). Un pdf reprenant ces precedents resultats. Et un tableau au format csv contenant le detail des alertes.
}
\description{
Une fonction permettant de determiner le nombre de point de prelevement par station et donnant des alertes dans le cas ou il y en aurait plusieurs.
}
\examples{
result_nombrePop <- test_sta_nombrePop(
data = exemple_IDF,
csv = FALSE,
pdf = FALSE)

}
